requirements.txt
pystun3==2.0.0
cachetools==5.5.1
redis==5.1.0